Fix inputcapture capability type for sending to KWin
Despite the name toInt() does not convert to int but to the underlying
type which is uint in this case. However the KWin D-Bus API takes a
signed int.

+++ src/inputcapture.cpp
@@ -76,7 +76,7 @@ InputCapturePortal::InputCapturePortal(QObject *parent)
 bool InputCapturePortal::setupInputCaptureSession(InputCaptureSession *session, Capabilities capabilities)
 {
     auto msg = QDBusMessage::createMethodCall(kwinService(), kwinInputCapturePath(), kwinInputCaptureManagerInterface(), u"addInputCapture"_s);
-    msg << capabilities.toInt();
+    msg << static_cast<int>(capabilities.toInt());
     QDBusReply<QDBusObjectPath> reply = QDBusConnection::sessionBus().call(msg, QDBus::Block, kwinDBusTimeout);
     if (!reply.isValid()) {
         qCWarning(XdgDesktopPortalKdeInputCapture) << "Failed to create KWin input capture:" << reply.error();
